+++ src/main/java/withboard/mvc/domain/MeetCategory.java
@@ -22,7 +22,7 @@ public class MeetCategory {
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "meet_category_no_seq")
 	@SequenceGenerator(sequenceName = "meet_category_no_seq", allocationSize = 1, name = "meet_category_no_seq")
-	private int meetCategoryNo;
+	private Long meetCategoryNo;
 	
 	@Column(nullable = false, length=100)
 	private String meetCategoryName;
